feat: add Pydantic models and enums
